Move `Dialog` to components folder for shared use
- This component is highly generic and will be useful across all applications.
- Additionally, I've added an icon for the close button and some CSS classes for future styling.
